Rename store folder to redux
